#1 Some server refactoring
