+++ package.json
@@ -7,8 +7,8 @@
     "yarn": "1.13.0"
   },
   "scripts": {
-    "start": "cross-env NODE_ENV=production babel-node server -r scripts/serverPreconfig.js",
-    "start:dev": "nodemon --watch server --watch webpack/dev.config.js --watch .env --exec 'cross-env NODE_ENV=development babel-node server -r scripts/serverPreconfig.js'",
+    "start": "cross-env NODE_ENV=production babel-node server",
+    "start:dev": "nodemon --watch server --watch webpack/dev.config.js --watch .env --exec 'cross-env NODE_ENV=development babel-node server'",
     "prebuild": "yarn build:clean",
     "build": "cross-env NODE_ENV=production webpack --config webpack/prod.config.js --color -p --progress --hide-modules --display-optimization-bailout",
     "build:dev": "cross-env NODE_ENV=development webpack --config webpack/dev.config.js --color -p --progress --hide-modules",
