+++ server/index.js
@@ -1,36 +1,11 @@
-import express from 'express';
-import { preloadAll } from 'react-loadable';
+require('dotenv').load();
 
-import setup from './setup';
-import getViewContext from './getViewContext';
-
-const app = express();
-
-setup(app);
-
-app.get('/*', async (req, res) => {
-  try {
-    const context = {};
-    const viewContext = await getViewContext(req, context);
-    if (context.url) {
-      res.redirect(context.url);
-      return;
-    }
-    res.render('index', viewContext);
-  } catch (e) {
-    console.error(e); // eslint-disable-line no-console
-    // TODO: implement custom server error page
-    res.send(500);
-  }
+require('@babel/register')({
+  presets: ['@babel/preset-env', '@babel/preset-react'],
+  plugins: ['dynamic-import-node'],
 });
 
-preloadAll()
-  .then(() => {
-    app.listen(process.env.PORT, () => {
-      // TODO: use logging middleware
-      console.log('Server has started'); // eslint-disable-line no-console
-    });
-  })
-  .catch(error => {
-    console.log(error); // eslint-disable-line no-console
-  });
+const ignore = require('ignore-styles');
+ignore.default([...ignore.DEFAULT_EXTENSIONS, '.ico', '.ttf']);
+
+require('./server');
