+++ webpack/loadable-stats-development.json
@@ -3189,7 +3189,7 @@
       "publicPath": "/bundles/front/commons.bundle-development.js"
     }
   ],
-  "../polyfills/extends": [
+  "../../polyfills/extends": [
     {
       "id": "./node_modules/emoji-mart/dist-es/polyfills/extends.js",
       "name": "./node_modules/emoji-mart/dist-es/polyfills/extends.js",
@@ -3197,7 +3197,7 @@
       "publicPath": "/bundles/front/commons.bundle-development.js"
     }
   ],
-  "../polyfills/inherits": [
+  "../../polyfills/inherits": [
     {
       "id": "./node_modules/emoji-mart/dist-es/polyfills/inherits.js",
       "name": "./node_modules/emoji-mart/dist-es/polyfills/inherits.js",
@@ -3205,7 +3205,7 @@
       "publicPath": "/bundles/front/commons.bundle-development.js"
     }
   ],
-  "../polyfills/objectGetPrototypeOf": [
+  "../../polyfills/objectGetPrototypeOf": [
     {
       "id": "./node_modules/emoji-mart/dist-es/polyfills/objectGetPrototypeOf.js",
       "name": "./node_modules/emoji-mart/dist-es/polyfills/objectGetPrototypeOf.js",
@@ -3213,7 +3213,7 @@
       "publicPath": "/bundles/front/commons.bundle-development.js"
     }
   ],
-  "../polyfills/possibleConstructorReturn": [
+  "../../polyfills/possibleConstructorReturn": [
     {
       "id": "./node_modules/emoji-mart/dist-es/polyfills/possibleConstructorReturn.js",
       "name": "./node_modules/emoji-mart/dist-es/polyfills/possibleConstructorReturn.js",
