+++ webpack/stats-production.json
@@ -1,8 +1,8 @@
 {
   "assetsByChunkName": {
-    "commons": "commons.4ad5133633c1cf7719e3.bundle-production.js",
-    "application": "application.4ad5133633c1cf7719e3.bundle-production.js",
-    "home-page": "home-page.4ad5133633c1cf7719e3.bundle-production.js",
-    "main": "main.4ad5133633c1cf7719e3.bundle-production.js"
+    "commons": "commons.70f01b33b9df20a4fadd.bundle-production.js",
+    "application": "application.70f01b33b9df20a4fadd.bundle-production.js",
+    "home-page": "home-page.70f01b33b9df20a4fadd.bundle-production.js",
+    "main": "main.70f01b33b9df20a4fadd.bundle-production.js"
   }
 }
